Separating express middleware from the server file
